requirements.txt
pip==22.0.2
attrs==24.3.0
wheel==0.37.1
PySocks==1.7.1
protobuf==3.12.4
cryptography==3.4.8
pytest==8.3.4
Pillow==9.0.1
keyring==23.5.0
Faker==33.3.0
typing_extensions==4.12.2
python-dateutil==2.8.1
pluggy==1.5.0
iniconfig==2.0.0
pexpect==4.8.0
certifi==2024.12.14
urllib3==1.26.5
six==1.16.0
idna==3.3
chardet==4.0.0
allure-python-commons==2.13.5
packaging==24.2
setuptools==59.6.0
future==0.18.2
allure-pytest==2.13.5
requests==2.25.1